requirements.txt
google-cloud-aiplatform==1.70.0
yt-dlp==2024.10.22
ffmpeg-python==0.2.0